型変更

+++ sql/4.6/orcadbup_ptbyomei_4.6.0-1.sql
@@ -1,4 +1,4 @@
 \set ON_ERROR_STOP
 
-alter table tbl_ptbyomei add column seqnum numeric(4) DEFAULT 0;
+alter table tbl_ptbyomei add column seqnum smallint DEFAULT 0;
 update tbl_ptbyomei set seqnum = 9999;
